moved ContainerHelper into PlayerUtil
