+++ src/main/java/teamport/aether/entity/MobUtil.java
@@ -4,7 +4,7 @@ import net.minecraft.core.entity.Entity;
 import net.minecraft.core.entity.Mob;
 import net.minecraft.core.entity.player.Player;
 import net.minecraft.core.util.helper.MathHelper;
-import teamport.aether.entity.player.ContainerHelper;
+import teamport.aether.entity.player.PlayerUntil;
 import teamport.aether.item.AetherArmorMaterial;
 
 public class MobUtil {
@@ -23,7 +23,7 @@ public class MobUtil {
         double distZ = attacker.z - target.z;
 
         if(target instanceof Player){
-            int count = ContainerHelper.countArmorPiecesOfMaterial(((Player)target).inventory, AetherArmorMaterial.OBSIDIAN);
+            int count = PlayerUntil.countArmorPiecesOfMaterial(((Player)target).inventory, AetherArmorMaterial.OBSIDIAN);
             if (count >= 5) {
                 return;
             }
