+++ src/main/java/teamport/aether/mixin/armor/player/neptune/MobMixinSwimming.java
@@ -12,7 +12,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Slice;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import teamport.aether.entity.player.ContainerHelper;
+import teamport.aether.entity.player.PlayerUntil;
 import teamport.aether.item.AetherArmorMaterial;
 
 @Mixin(value = Mob.class, remap = false)
@@ -26,7 +26,7 @@ public abstract class MobMixinSwimming extends Entity {
             return;
         }
         Player player = (Player) (Object) this;
-        if (ContainerHelper.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.NEPTUNE) < 5) {
+        if (PlayerUntil.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.NEPTUNE) < 5) {
             return;
         }
         yd += 0.02;
@@ -40,7 +40,7 @@ public abstract class MobMixinSwimming extends Entity {
         }
         Player player = (Player) (Object) this;
 
-        if (ContainerHelper.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.NEPTUNE) < 5) {
+        if (PlayerUntil.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.NEPTUNE) < 5) {
             return constant;
         }
         return this.speed * 0.4f;
@@ -52,7 +52,7 @@ public abstract class MobMixinSwimming extends Entity {
         }
         Player player = (Player) (Object) this;
 
-        if (ContainerHelper.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.NEPTUNE) < 5) {
+        if (PlayerUntil.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.NEPTUNE) < 5) {
             return constant;
         }
         return 0.16;
