+++ src/main/java/teamport/aether/mixin/armor/player/obsidian/PlayerMixinBlastImmunity.java
@@ -8,7 +8,7 @@ import net.minecraft.core.util.helper.DamageType;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
-import teamport.aether.entity.player.ContainerHelper;
+import teamport.aether.entity.player.PlayerUntil;
 import teamport.aether.item.AetherArmorMaterial;
 
 @Mixin(value = Player.class, remap = false)
@@ -17,7 +17,7 @@ public abstract class PlayerMixinBlastImmunity {
     public ContainerInventory inventory;
     @ModifyExpressionValue(method = "hurt", at = @At(value = "INVOKE", target = "Lnet/minecraft/core/entity/Mob;hurt(Lnet/minecraft/core/entity/Entity;ILnet/minecraft/core/util/helper/DamageType;)Z"))
     private boolean negateDamage(boolean original, Entity attacker, int damage, DamageType type) {
-        if (type == null || type.equals(DamageType.BLAST) || ContainerHelper.countArmorPiecesOfMaterial(this.inventory, AetherArmorMaterial.OBSIDIAN) < 5) {
+        if (type == null || type.equals(DamageType.BLAST) || PlayerUntil.countArmorPiecesOfMaterial(this.inventory, AetherArmorMaterial.OBSIDIAN) < 5) {
             return original;
         }
         this.inventory.damageArmor((int) Math.ceil((double) damage / (double) 4.0F));
