+++ src/main/java/teamport/aether/mixin/armor/player/obsidian/PlayerMixinNoKnockbackFling.java
@@ -5,7 +5,7 @@ import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import net.minecraft.core.entity.Entity;
 import net.minecraft.core.entity.player.Player;
 import org.spongepowered.asm.mixin.Mixin;
-import teamport.aether.entity.player.ContainerHelper;
+import teamport.aether.entity.player.PlayerUntil;
 import teamport.aether.item.AetherArmorMaterial;
 
 @Mixin(value = Entity.class, remap = false)
@@ -17,10 +17,10 @@ public abstract class PlayerMixinNoKnockbackFling {
             return;
         }
         Player player = (Player) (Object) this;
-        if (ContainerHelper.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.OBSIDIAN) >= 5) {
+        if (PlayerUntil.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.OBSIDIAN) >= 5) {
             return;
         }
-        if (ContainerHelper.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.OBSIDIAN) >= 3) {
+        if (PlayerUntil.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.OBSIDIAN) >= 3) {
             original.call(xd / 4, yd / 4, zd / 4, pushTime / 4);
             return;
         }
