+++ src/main/java/teamport/aether/mixin/armor/player/obsidian/PlayerMixinNoKnockbackKnockBack.java
@@ -6,7 +6,7 @@ import net.minecraft.core.entity.Entity;
 import net.minecraft.core.entity.Mob;
 import net.minecraft.core.entity.player.Player;
 import org.spongepowered.asm.mixin.Mixin;
-import teamport.aether.entity.player.ContainerHelper;
+import teamport.aether.entity.player.PlayerUntil;
 import teamport.aether.item.AetherArmorMaterial;
 
 @Mixin(value = Mob.class, remap = false)
@@ -18,10 +18,10 @@ public abstract class PlayerMixinNoKnockbackKnockBack {
             return;
         }
         Player player = (Player) (Object) this;
-        if (ContainerHelper.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.OBSIDIAN) >= 5) {
+        if (PlayerUntil.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.OBSIDIAN) >= 5) {
             return;
         }
-        if (ContainerHelper.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.OBSIDIAN) >= 3) {
+        if (PlayerUntil.countArmorPiecesOfMaterial(player.inventory, AetherArmorMaterial.OBSIDIAN) >= 3) {
             original.call(entity, i, d / 2, d1 / 2);
             return;
         }
